Move mojos to dedicated package
