+++ heroku-maven-plugin/src/main/java/com/heroku/sdk/maven/mojo/DeployOnlyMojo.java
@@ -1,11 +1,8 @@
-package com.heroku.sdk.maven;
-
-import java.util.Map;
+package com.heroku.sdk.maven.mojo;
 
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Mojo;
-import org.apache.maven.plugins.annotations.Parameter;
 
 /**
  * Deploys an application to Heroku without resolving dependencies, packaging, or vendoring dependencies. This
